+++ docs/templates/cjs/authentication.js
@@ -1,6 +1,5 @@
 const UserModel = require('../models/UserModel');
 
-
 /**
  * User authentication configuration interface.
  */
@@ -16,6 +15,15 @@ module.exports = {
    * @type {'memory|redis'}
    */
   session_store: 'memory',
+
+  /**
+   * If the session is stored in "redis", this field is required and should be set to the hostname of the Redis server.
+   * For example, to connect to redis on localhost on port 6379, set "redis://localhost:6379".
+   * To connect to a different host or port, use a connection string in the format "redis[s]://[[username][:password]@][host][:port][/db-number]".
+   * For example, "redis://alice:foobared@awesome.redis.server:6380".
+   * @type {string|undefined}
+   */
+  redis_host: undefined,
 
   /**
    * Authentication user ID field name, defaults to `username`.