+++ src/interfaces/AuthenticationOptions.ts
@@ -17,6 +17,15 @@ export default interface {
   session_store: 'memory'|'redis',
 
   /**
+   * If the session is stored in "redis", this field is required and should be set to the hostname of the Redis server.
+   * For example, to connect to redis on localhost on port 6379, set "redis://localhost:6379".
+   * To connect to a different host or port, use a connection string in the format "redis[s]://[[username][:password]@][host][:port][/db-number]".
+   * For example, "redis://alice:foobared@awesome.redis.server:6380".
+   * @type {string|undefined}
+   */
+  redis_host?: string,
+
+  /**
    * Authentication user ID field name, defaults to `username`.
    * @type {string}
    */