+++ types/middlewares/CORS.d.ts
@@ -6,7 +6,7 @@ export default class {
     /**
      * Mount on application.
      */
-    static mount(app: express.Express): undefined;
+    static mount(app: express.Express): void;
     /**
      * Returns the option.
      *